Refactored the directory (Courtesy of EternalSyntaxError)
